requirements.txt
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-rest-registration==0.9.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
sqlparse==0.5.3
pytz==2025.1
tzdata==2025.1
